feat: community-page 라우터 설정
