+++ apps/client/src/shared/router/routes/global-routes.tsx
@@ -1,4 +1,6 @@
-import CommunityPage from '@pages/community/community-page.tsx';
+import CommunityPage from '@pages/community/community-page';
+import CommunityPostWrite from '@pages/community/community-post-write/community-post-write';
+import CommunityWrite from '@pages/community/community-write/community-write';
 import LoginPage from '@pages/login/login-page.tsx';
 import LoginFallbackPage from '@pages/login-fallback/login-fallback-page.tsx';
 import MyPage from '@pages/my/my-page.tsx';
@@ -39,7 +41,8 @@ export const protectedRoutes = [
     path: routePath.COMMUNITY,
     Component: CommunityPage,
     children: [
-      // @TODO 커뮤니티 관련 하위 라우트 작성
+      { path: 'post', Component: CommunityPostWrite },
+      { path: 'write', Component: CommunityWrite },
     ],
   },
   {
